note-app-v4
Authentication & Authorization

+++ note-app-v4/src/api/notes/handler.js
@@ -16,8 +16,9 @@ class NotesHandler{
         try {
             this._validator.validateNotePayload(req.payload)
             const {title , tags, body} = req.payload 
+            const {id : credentialId} = req.auth.credentials
 
-            const noteId = await this._service.addNote({title, tags, body})
+            const noteId = await this._service.addNote({title, tags, body, owner: credentialId})
     
             const response = h.response({
                 status : 'success',
@@ -49,8 +50,9 @@ class NotesHandler{
         }
     }
 
-    async getNotesHandler(){
-        const notes = await this._service.getNotes()
+    async getNotesHandler(req){
+        const {id: credentialId} = req.auth.credentials
+        const notes = await this._service.getNotes(credentialId)
         return {
             status : 'success',
             data : {
@@ -62,6 +64,9 @@ class NotesHandler{
     async getNoteByIdHandler(req, h){
         try {
             const {id} = req.params
+            const {id : credentialId} = req.auth.credentials
+
+            await this._service.verifyNoteOwner(id, credentialId)
             const note = await this._service.getNoteById(id)
 
             return {
@@ -95,7 +100,9 @@ class NotesHandler{
         try {
             this._validator.validateNotePayload(req.payload)
             const {id} = req.params
+            const {id : credentialId} = req.auth.credentials
 
+            await this._service.verifyNoteOwner(id, credentialId)
             await this._service.editNoteById(id, req.payload)
             
             return {
@@ -126,7 +133,9 @@ class NotesHandler{
     async deleteNoteByIdHandler(req, h){
         try {
             const {id} = req.params
+            const {id: credentialId} = req.auth.credentials
 
+            await this._service.verifyNoteOwner(id, credentialId)
             await this._service.deleteNoteById(id)
 
             return {
